Custom grouping works now
