+++ heron/instance/src/python/basics/bolt.py
@@ -142,8 +142,7 @@ class Bolt(Component):
     if custom_target_task_ids is not None:
       for task_id in custom_target_task_ids:
         # for custom grouping
-        to_add = data_tuple.dest_task_ids.add()
-        to_add = task_id
+        data_tuple.dest_task_ids.append(task_id)
 
     # Set the anchors for a tuple
     if anchors is not None:
